requirements.txt
fastapi
uvicorn[standard]
redis
rq
httpx
python-dotenv
pydantic
pydantic-settings
google-genai
